Fix issue : Delete likes of a group when deleting it

+++ classes/components/GroupsComponent.php
@@ -708,6 +708,10 @@ class GroupsComponent {
 	 */
 	public function delete_group(int $groupID) : bool {
 
+		// Delete all the likes of the group
+		if(!components()->likes->delete_all($groupID, Likes::LIKE_GROUP))
+			return FALSE;
+
 		//Delete group image
 		if(!$this->deleteLogo($groupID))
 			return FALSE;
